+++ packages/hextorgb/src/typings.ts
@@ -1,0 +1,5 @@
+export type RGBColor = {
+  r: number;
+  g: number;
+  b: number;
+};