+++ packages/media/src/utils.ts
@@ -1,6 +1,4 @@
-export function isString(arg: unknown): boolean {
-  return Object.prototype.toString.call(arg) === '[object String]';
-}
+import { isString } from '@annexe/isstring';
 
 export function isNonEmptyString(arg?: unknown): boolean {
   return isString(arg) && (arg as string).trim() !== '';