+++ packages/pxtoem/src/index.ts
@@ -1,0 +1,5 @@
+export {
+  pxToEm,
+  getPxNumericValue,
+  extractPxNumericValue
+} from '../src/pxtoem';