+++ packages/imagery/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @artifak/image
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [6f08e4c]
+  - @artifak/component-generator@2.0.3
+
 ## 1.0.7
 
 ### Patch Changes